% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DVLA.R
\name{getDVLA}
\alias{getDVLA}
\title{Returns a data frame of vehicle data from the DVLA about a single vehicle}
\usage{
getDVLA(key, reg)
}
\arguments{
\item{key}{The API key provided to you by the DVLA}

\item{reg}{A single character containing one vehicle registration number}
}
\description{
\code{getDVLA}:getDVLA takes a vehicle registration plate number and returns
information held by the DVLA about this vehicle. To use this service you
must have registered and been granted an API key.
}
