% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DVLA.R
\name{tidyDVLA}
\alias{tidyDVLA}
\title{Tidy version of getDVLA for a data frame containing vehicle registrations}
\usage{
tidyDVLA(data, reg_column, key)
}
\arguments{
\item{data}{A data frame containing a vehicle registration column}

\item{reg_column}{The quoted column name that contains vehicle registrations}

\item{key}{The API key provided to you by the DVLA}
}
\description{
\code{tidyDVLA}: tidyDVLA extends the functionality of \code{getDVLA} by
taking a data frame containing vehicle registration numbers and using
parallel processing to retrieve DVLA information for each one. It binds
these columns back to the input data frame.
}
